requirements.txt
pandas
python-dotenv
tqdm
sqlparse
pydantic